+++ doc/CHANGES.rst
@@ -14,6 +14,10 @@ CHANGELOG
 Future
 ------
 
+* [FIX] Allow to add a relationship on the same model, the main use is to add 
+  parent / children relation ship on a model, They are any difference with
+  the declaration of ta relation ship on another model
+
 0.2.2
 -----